BootCamp-Python/JdD_011/pyproject.toml
[tool.poetry]
name = "jdd-011"
version = "0.1.0"
description = ""
authors = ["Durso <dursoarthur@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
